add goalTabPost toggle button and ux improvement
