+++ components/GoalTabPostToggleButton.vue
@@ -1,0 +1,11 @@
+<template>
+  <v-btn size="small" @click="$emit('onAddGoalButtonClick')">
+    <v-icon start icon="mdi:mdi-plus" />
+    Add Goal
+  </v-btn>
+</template>
+
+<script lang="ts" setup>
+</script>
+
+<style scoped></style>
