+++ pages/index.vue
@@ -1,29 +1,19 @@
 <template>
   <v-container>
-    <GoalTabAdd :goal="DEFAULT_GOAL" @goal-added="addGoal" />
-    <v-btn @click="showGoalForm = true" color="primary">Show Goal Modal</v-btn>
-    <GoalFormModal :visible="showGoalForm" @close="showGoalForm = false" />
+    <GoalTabPost v-if="showGoalTabPost" @on-goal-tab-post-blur="showGoalTabPost = false"/>
+    <GoalTabPostToggleButton v-else @on-add-goal-button-click="showGoalTabPost = true"/>
   </v-container>
 </template>
 
 <script>
-import GoalFormModal from '~/components/GoalFormModal.vue'
-import GoalTabAdd from '~/components/GoalTabAdd.vue'
-import { DEFAULT_GOAL } from '~/constants/goal'
-
 export default {
   name: 'HomePage',
   data () {
     return {
-      DEFAULT_GOAL,
-      showGoalForm: false
+      showGoalForm: false,
+      showGoalTabPost: false
     }
   },
-  methods: {
-    addGoal (goal) {
-    }
-  },
-  components: { GoalTabAdd, GoalFormModal }
 }
 </script>
